+++ Changelog.md
@@ -8,7 +8,9 @@ CHANGELOG
 
 ### 3.2.0-preview
 
-- WinCopies.GUI.Windows.Window's Close Win32 system command can be disabled using dependency property.
+- WinCopies.GUI.Windows.Window:
+	- Close Win32 system command can be disabled using dependency property.
+	- ShowHelpButton is now HelpButton
 - ProcessWindow inherits from WinCopies.GUI.Windows.Window.
 
 ### 04/26/2021 3.1.0-preview